refactor: reorganize codegen/mir/parser modules and move parser tests
