--- src/codegen/cranelift/mod.rs
@@ -1,9 +1,0 @@
-pub mod clif_builder;
-pub mod jit;
-pub mod metadata;
-pub(crate) mod mir_parse;
-pub mod object_builder;
-pub mod ometa;
-pub(crate) mod parse_utils;
-pub mod type_layout;
-pub mod typed_codegen;
