+++ src/codegen/typed/mod.rs
@@ -1,0 +1,9 @@
+mod control_flow;
+mod helpers;
+mod lowering;
+mod shadow;
+mod signatures;
+mod types;
+
+pub use lowering::define_typed_function;
+pub(crate) use signatures::function_signature;
