+++ src/mir/typed/mod.rs
@@ -1,0 +1,13 @@
+include!("shared.rs");
+
+mod builder;
+mod stmt_lowering;
+mod expr_lowering;
+mod control_flow;
+
+use builder::*;
+use control_flow::*;
+use expr_lowering::*;
+use stmt_lowering::*;
+
+pub use builder::typed;
