+++ src/tests/mod.rs
@@ -1,0 +1,5 @@
+mod codegen_jit;
+mod codegen_object_builder;
+mod codegen_ometa;
+mod mir_typed;
+mod parser;
